+++ dev/src/serializer.ts
@@ -118,7 +118,8 @@ export class Serializer {
     }
 
     if (typeof val === 'number') {
-      if (Number.isSafeInteger(val)) {
+      const isNegativeZero = val === 0 && 1 / val === 1 / -0;
+      if (Number.isSafeInteger(val) && !isNegativeZero) {
         return {
           integerValue: val as number,
         };