+++ dev/test/document.ts
@@ -224,6 +224,29 @@ describe('serialize document', () => {
     });
   });
 
+  it('serializes negative zero into double', () => {
+    const overrides: ApiOverride = {
+      commit: request => {
+        requestEquals(
+          request,
+          set({
+            document: document('documentId', 'negativeZero', {
+              doubleValue: -0,
+            }),
+          })
+        );
+        return response(writeResult(1));
+      },
+    };
+
+    return createInstance(overrides).then(firestore => {
+      return firestore.doc('collectionId/documentId').set({
+        // Set to -0, which should be stored as a double.
+        negativeZero: -0,
+      });
+    });
+  });
+
   it('serializes date before 1970', () => {
     const overrides: ApiOverride = {
       commit: request => {
@@ -1309,47 +1332,6 @@ describe('set document', () => {
 
   it('validates merge option', () => {
     expect(() => {
-      firestore
-        .doc('collectionId/documentId')
-        .set({foo: 'bar'}, 'foo' as InvalidApiUsage);
-    }).to.throw(
-      'Value for argument "options" is not a valid set() options argument. Input is not an object.'
-    );
-
-    expect(() => {
-      firestore.doc('collectionId/documentId').set(
-        {foo: 'bar'},
-        {
-          merge: 42 as InvalidApiUsage,
-        }
-      );
-    }).to.throw(
-      'Value for argument "options" is not a valid set() options argument. "merge" is not a boolean.'
-    );
-
-    expect(() => {
-      firestore.doc('collectionId/documentId').set(
-        {foo: 'bar'},
-        {
-          mergeFields: 42 as InvalidApiUsage,
-        }
-      );
-    }).to.throw(
-      'Value for argument "options" is not a valid set() options argument. "mergeFields" is not an array.'
-    );
-
-    expect(() => {
-      firestore.doc('collectionId/documentId').set(
-        {foo: 'bar'},
-        {
-          mergeFields: [null as InvalidApiUsage],
-        }
-      );
-    }).to.throw(
-      'Value for argument "options" is not a valid set() options argument. "mergeFields" is not valid: Element at index 0 is not a valid field path. Paths can only be specified as strings or via a FieldPath object.'
-    );
-
-    expect(() => {
       firestore.doc('collectionId/documentId').set(
         {foo: 'bar'},
         {
@@ -1357,6 +1339,19 @@ describe('set document', () => {
         }
       );
     }).to.throw('Input data is missing for field "foobar".');
+
+    expect(() => {
+      firestore.doc('collectionId/documentId').set(
+        {foo: 'bar'},
+        {
+          mergeFields: ['foobar..'],
+        }
+      );
+    }).to.throw(
+      'Value for argument "options" is not a valid set() options argument. ' +
+        '"mergeFields" is not valid: Element at index 0 is not a valid ' +
+        'field path. Paths must not contain ".." in them.'
+    );
 
     expect(() => {
       firestore