+++ types/firestore.d.ts
@@ -28,9 +28,34 @@ declare namespace FirebaseFirestore {
   export type DocumentData = {[field: string]: any};
 
   /**
+   * Similar to Typescript's `Partial<T>`, but allows nested fields to be
+   * omitted and FieldValues to be passed in as property values.
+   */
+  export type PartialWithFieldValue<T> =
+    | Partial<T>
+    | (T extends Primitive
+        ? T
+        : T extends {}
+        ? {[K in keyof T]?: PartialWithFieldValue<T[K]> | FieldValue}
+        : never);
+
+  /**
+   * Allows FieldValues to be passed in as a property value while maintaining
+   * type safety.
+   */
+  export type WithFieldValue<T> =
+    | T
+    | (T extends Primitive
+        ? T
+        : T extends {}
+        ? {[K in keyof T]: WithFieldValue<T[K]> | FieldValue}
+        : never);
+
+  /**
    * Update data (for use with [update]{@link DocumentReference#update})
    * that contains paths mapped to values. Fields that contain dots reference
-   * nested fields within the document.
+   * nested fields within the document. FieldValues can be passed in
+   * as property values.
    *
    * You can update a top-level field in your document by using the field name
    * as a key (e.g. `foo`). The provided value completely replaces the contents
@@ -40,7 +65,72 @@ declare namespace FirebaseFirestore {
    * key (e.g. `foo.bar`). This nested field update replaces the contents at
    * `bar` but does not modify other data under `foo`.
    */
-  export type UpdateData = {[fieldPath: string]: any};
+  export type UpdateData<T> = T extends Primitive
+    ? T
+    : T extends {}
+    ? {[K in keyof T]?: UpdateData<T[K]> | FieldValue} & NestedUpdateFields<T>
+    : Partial<T>;
+
+  /** Primitive types. */
+  export type Primitive = string | number | boolean | undefined | null;
+
+  /**
+   * For each field (e.g. 'bar'), find all nested keys (e.g. {'bar.baz': T1,
+   * 'bar.qux': T2}). Intersect them together to make a single map containing
+   * all possible keys that are all marked as optional
+   */
+  export type NestedUpdateFields<T extends Record<string, unknown>> =
+    UnionToIntersection<
+      {
+        [K in keyof T & string]: ChildUpdateFields<K, T[K]>;
+      }[keyof T & string] // Also include the generated prefix-string keys.
+    >;
+
+  /**
+   * Helper for calculating the nested fields for a given type T1. This is needed
+   * to distribute union types such as `undefined | {...}` (happens for optional
+   * props) or `{a: A} | {b: B}`.
+   *
+   * In this use case, `V` is used to distribute the union types of `T[K]` on
+   * `Record`, since `T[K]` is evaluated as an expression and not distributed.
+   *
+   * See https://www.typescriptlang.org/docs/handbook/advanced-types.html#distributive-conditional-types
+   */
+  export type ChildUpdateFields<K extends string, V> =
+    // Only allow nesting for map values
+    V extends Record<string, unknown>
+      ? // Recurse into the map and add the prefix in front of each key
+        // (e.g. Prefix 'bar.' to create: 'bar.baz' and 'bar.qux'.
+        AddPrefixToKeys<K, UpdateData<V>>
+      : // UpdateData is always a map of values.
+        never;
+
+  /**
+   * Returns a new map where every key is prefixed with the outer key appended
+   * to a dot.
+   */
+  export type AddPrefixToKeys<
+    Prefix extends string,
+    T extends Record<string, unknown>
+  > =
+    // Remap K => Prefix.K. See https://www.typescriptlang.org/docs/handbook/2/mapped-types.html#key-remapping-via-as
+    {[K in keyof T & string as `${Prefix}.${K}`]+?: T[K]};
+
+  /**
+   * Given a union type `U = T1 | T2 | ...`, returns an intersected type
+   * `(T1 & T2 & ...)`.
+   *
+   * Uses distributive conditional types and inference from conditional types.
+   * This works because multiple candidates for the same type variable in
+   * contra-variant positions causes an intersection type to be inferred.
+   * https://www.typescriptlang.org/docs/handbook/advanced-types.html#type-inference-in-conditional-types
+   * https://stackoverflow.com/questions/50374908/transform-union-type-to-intersection-type
+   */
+  export type UnionToIntersection<U> = (
+    U extends unknown ? (k: U) => void : never
+  ) extends (k: infer I) => void
+    ? I
+    : never;
 
   /**
    * Sets or disables the log function for all active Firestore instances.
@@ -95,9 +185,27 @@ declare namespace FirebaseFirestore {
      * into a plain Javascript object (suitable for writing directly to the
      * Firestore database). To use set() with `merge` and `mergeFields`,
      * toFirestore() must be defined with `Partial<T>`.
+     *
+     * The `WithFieldValue<T>` type extends `T` to also allow FieldValues such
+     * as `FieldValue.delete()` to be used as property values.
      */
-    toFirestore(modelObject: T): DocumentData;
-    toFirestore(modelObject: Partial<T>, options: SetOptions): DocumentData;
+    toFirestore(modelObject: WithFieldValue<T>): DocumentData;
+
+    /**
+     * Called by the Firestore SDK to convert a custom model object of type T
+     * into a plain Javascript object (suitable for writing directly to the
+     * Firestore database). To use set() with `merge` and `mergeFields`,
+     * toFirestore() must be defined with `Partial<T>`.
+     *
+     * The `PartialWithFieldValue<T>` type extends `Partial<T>` to allow
+     * FieldValues such as `FieldValue.delete()` to be used as property values.
+     * It also supports nested `Partial` by allowing nested fields to be
+     * omitted.
+     */
+    toFirestore(
+      modelObject: PartialWithFieldValue<T>,
+      options: SetOptions
+    ): DocumentData;
 
     /**
      * Called by the Firestore SDK to convert Firestore data into an object of
@@ -490,9 +598,13 @@ declare namespace FirebaseFirestore {
      *
      * @param documentRef A reference to the document to be create.
      * @param data The object data to serialize as the document.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return This `Transaction` instance. Used for chaining method calls.
      */
-    create<T>(documentRef: DocumentReference<T>, data: T): Transaction;
+    create<T>(
+      documentRef: DocumentReference<T>,
+      data: WithFieldValue<T>
+    ): Transaction;
 
     /**
      * Writes to the document referred to by the provided `DocumentReference`.
@@ -502,14 +614,26 @@ declare namespace FirebaseFirestore {
      * @param documentRef A reference to the document to be set.
      * @param data An object of the fields and values for the document.
      * @param options An object to configure the set behavior.
+     * @param  options.merge - If true, set() merges the values specified in its
+     * data argument. Fields omitted from this set() call remain untouched. If
+     * your input sets any field to an empty map, all nested fields are
+     * overwritten.
+     * @param options.mergeFields - If provided, set() only replaces the
+     * specified field paths. Any field path that is not specified is ignored
+     * and remains untouched. If your input sets any field to an empty map, all
+     * nested fields are overwritten.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return This `Transaction` instance. Used for chaining method calls.
      */
     set<T>(
       documentRef: DocumentReference<T>,
-      data: Partial<T>,
+      data: PartialWithFieldValue<T>,
       options: SetOptions
     ): Transaction;
-    set<T>(documentRef: DocumentReference<T>, data: T): Transaction;
+    set<T>(
+      documentRef: DocumentReference<T>,
+      data: WithFieldValue<T>
+    ): Transaction;
 
     /**
      * Updates fields in the document referred to by the provided
@@ -523,11 +647,12 @@ declare namespace FirebaseFirestore {
      * @param data An object containing the fields and values with which to
      * update the document.
      * @param precondition A Precondition to enforce on this update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @return This `Transaction` instance. Used for chaining method calls.
      */
-    update(
-      documentRef: DocumentReference<any>,
-      data: UpdateData,
+    update<T>(
+      documentRef: DocumentReference<T>,
+      data: UpdateData<T>,
       precondition?: Precondition
     ): Transaction;
 
@@ -548,6 +673,7 @@ declare namespace FirebaseFirestore {
      * @param fieldsOrPrecondition An alternating list of field paths and values
      * to update, optionally followed by a `Precondition` to enforce on this
      * update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @return This `Transaction` instance. Used for chaining method calls.
      */
     update(
@@ -586,11 +712,15 @@ declare namespace FirebaseFirestore {
      * @param documentRef A reference to the document to be
      * created.
      * @param data The object to serialize as the document.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @returns A promise that resolves with the result of the write. If the
      * write fails, the promise is rejected with a
      * [BulkWriterError]{@link BulkWriterError}.
      */
-    create<T>(documentRef: DocumentReference<T>, data: T): Promise<WriteResult>;
+    create<T>(
+      documentRef: DocumentReference<T>,
+      data: WithFieldValue<T>
+    ): Promise<WriteResult>;
 
     /**
      * Delete a document from the database.
@@ -624,22 +754,28 @@ declare namespace FirebaseFirestore {
      * set.
      * @param data The object to serialize as the document.
      * @param options An object to configure the set behavior.
-     * @param  options.merge - If true, set() merges the values
-     * specified in its data argument. Fields omitted from this set() call
-     * remain untouched.
-     * @param options.mergeFields - If provided,
-     * set() only replaces the specified field paths. Any field path that is not
-     * specified is ignored and remains untouched.
+     * @param  options.merge - If true, set() merges the values specified in its
+     * data argument. Fields omitted from this set() call remain untouched. If
+     * your input sets any field to an empty map, all nested fields are
+     * overwritten.
+     * @param options.mergeFields - If provided, set() only replaces the
+     * specified field paths. Any field path that is not specified is ignored
+     * and remains untouched. If your input sets any field to an empty map, all
+     * nested fields are overwritten.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @returns A promise that resolves with the result of the write. If the
      * write fails, the promise is rejected with a
      * [BulkWriterError]{@link BulkWriterError}.
      */
     set<T>(
       documentRef: DocumentReference<T>,
-      data: Partial<T>,
+      data: PartialWithFieldValue<T>,
       options: SetOptions
     ): Promise<WriteResult>;
-    set<T>(documentRef: DocumentReference<T>, data: T): Promise<WriteResult>;
+    set<T>(
+      documentRef: DocumentReference<T>,
+      data: WithFieldValue<T>
+    ): Promise<WriteResult>;
 
     /**
      * Update fields of the document referred to by the provided
@@ -660,13 +796,14 @@ declare namespace FirebaseFirestore {
      * @param data An object containing the fields and values with which to
      * update the document.
      * @param precondition A Precondition to enforce on this update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @returns A promise that resolves with the result of the write. If the
      * write fails, the promise is rejected with a
      * [BulkWriterError]{@link BulkWriterError}.
      */
-    update(
-      documentRef: DocumentReference<any>,
-      data: UpdateData,
+    update<T>(
+      documentRef: DocumentReference<T>,
+      data: UpdateData<T>,
       precondition?: Precondition
     ): Promise<WriteResult>;
 
@@ -689,7 +826,9 @@ declare namespace FirebaseFirestore {
      * @param field The first field to update.
      * @param value The first value
      * @param fieldsOrPrecondition An alternating list of field paths and values
-     * to update, optionally followed a `Precondition` to enforce on this update.
+     * to update, optionally followed a `Precondition` to enforce on this
+     * update.
+     * @throws Error If the provided input is not valid Firestore data;
      * @returns A promise that resolves with the result of the write. If the
      * write fails, the promise is rejected with a
      * [BulkWriterError]{@link BulkWriterError}.
@@ -833,9 +972,13 @@ declare namespace FirebaseFirestore {
      *
      * @param documentRef A reference to the document to be created.
      * @param data The object data to serialize as the document.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return This `WriteBatch` instance. Used for chaining method calls.
      */
-    create<T>(documentRef: DocumentReference<T>, data: T): WriteBatch;
+    create<T>(
+      documentRef: DocumentReference<T>,
+      data: WithFieldValue<T>
+    ): WriteBatch;
 
     /**
      * Write to the document referred to by the provided `DocumentReference`.
@@ -845,14 +988,26 @@ declare namespace FirebaseFirestore {
      * @param documentRef A reference to the document to be set.
      * @param data An object of the fields and values for the document.
      * @param options An object to configure the set behavior.
+     * @param  options.merge - If true, set() merges the values specified in its
+     * data argument. Fields omitted from this set() call remain untouched. If
+     * your input sets any field to an empty map, all nested fields are
+     * overwritten.
+     * @param options.mergeFields - If provided, set() only replaces the
+     * specified field paths. Any field path that is not specified is ignored
+     * and remains untouched. If your input sets any field to an empty map, all
+     * nested fields are overwritten.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return This `WriteBatch` instance. Used for chaining method calls.
      */
     set<T>(
       documentRef: DocumentReference<T>,
-      data: Partial<T>,
+      data: PartialWithFieldValue<T>,
       options: SetOptions
     ): WriteBatch;
-    set<T>(documentRef: DocumentReference<T>, data: T): WriteBatch;
+    set<T>(
+      documentRef: DocumentReference<T>,
+      data: WithFieldValue<T>
+    ): WriteBatch;
 
     /**
      * Update fields of the document referred to by the provided
@@ -866,11 +1021,12 @@ declare namespace FirebaseFirestore {
      * @param data An object containing the fields and values with which to
      * update the document.
      * @param precondition A Precondition to enforce on this update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @return This `WriteBatch` instance. Used for chaining method calls.
      */
-    update(
-      documentRef: DocumentReference<any>,
-      data: UpdateData,
+    update<T>(
+      documentRef: DocumentReference<T>,
+      data: UpdateData<T>,
       precondition?: Precondition
     ): WriteBatch;
 
@@ -889,7 +1045,9 @@ declare namespace FirebaseFirestore {
      * @param field The first field to update.
      * @param value The first value
      * @param fieldsOrPrecondition An alternating list of field paths and values
-     * to update, optionally followed a `Precondition` to enforce on this update.
+     * to update, optionally followed a `Precondition` to enforce on this
+     * update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @return This `WriteBatch` instance. Used for chaining method calls.
      */
     update(
@@ -943,25 +1101,24 @@ declare namespace FirebaseFirestore {
    * `DocumentReference`, `WriteBatch` and `Transaction`. These calls can be
    * configured to perform granular merges instead of overwriting the target
    * documents in their entirety.
+   *
+   * @param merge Changes the behavior of a set() call to only replace the
+   * values specified in its data argument. Fields omitted from the set() call
+   * remain untouched. If your input sets any field to an empty map, all nested
+   * fields are overwritten.
+   *
+   * @param mergeFields Changes the behavior of set() calls to only replace
+   * the specified field paths. Any field path that is not specified is ignored
+   * and remains untouched. If your input sets any field to an empty map, all
+   * nested fields are overwritten.
    */
-  export interface SetOptions {
-    /**
-     * Changes the behavior of a set() call to only replace the values specified
-     * in its data argument. Fields omitted from the set() call remain
-     * untouched.
-     */
-    readonly merge?: boolean;
-
-    /**
-     * Changes the behavior of set() calls to only replace the specified field
-     * paths. Any field path that is not specified is ignored and remains
-     * untouched.
-     *
-     * It is an error to pass a SetOptions object to a set() call that is
-     * missing a value for any of the fields specified here.
-     */
-    readonly mergeFields?: (string | FieldPath)[];
-  }
+  export type SetOptions =
+    | {
+        readonly merge?: boolean;
+      }
+    | {
+        readonly mergeFields?: Array<string | FieldPath>;
+      };
 
   /**
    * An options object that can be used to configure the behavior of `getAll()`
@@ -1051,9 +1208,10 @@ declare namespace FirebaseFirestore {
      * provided object values. The write fails if the document already exists
      *
      * @param data The object data to serialize as the document.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return A Promise resolved with the write time of this create.
      */
-    create(data: T): Promise<WriteResult>;
+    create(data: WithFieldValue<T>): Promise<WriteResult>;
 
     /**
      * Writes to the document referred to by this `DocumentReference`. If the
@@ -1062,10 +1220,22 @@ declare namespace FirebaseFirestore {
      *
      * @param data A map of the fields and values for the document.
      * @param options An object to configure the set behavior.
+     * @param  options.merge - If true, set() merges the values specified in its
+     * data argument. Fields omitted from this set() call remain untouched. If
+     * your input sets any field to an empty map, all nested fields are
+     * overwritten.
+     * @param options.mergeFields - If provided, set() only replaces the
+     * specified field paths. Any field path that is not specified is ignored
+     * and remains untouched. If your input sets any field to an empty map, all
+     * nested fields are overwritten.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return A Promise resolved with the write time of this set.
      */
-    set(data: Partial<T>, options: SetOptions): Promise<WriteResult>;
-    set(data: T): Promise<WriteResult>;
+    set(
+      data: PartialWithFieldValue<T>,
+      options: SetOptions
+    ): Promise<WriteResult>;
+    set(data: WithFieldValue<T>): Promise<WriteResult>;
 
     /**
      * Updates fields in the document referred to by this `DocumentReference`.
@@ -1077,9 +1247,13 @@ declare namespace FirebaseFirestore {
      * @param data An object containing the fields and values with which to
      * update the document.
      * @param precondition A Precondition to enforce on this update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @return A Promise resolved with the write time of this update.
      */
-    update(data: UpdateData, precondition?: Precondition): Promise<WriteResult>;
+    update(
+      data: UpdateData<T>,
+      precondition?: Precondition
+    ): Promise<WriteResult>;
 
     /**
      * Updates fields in the document referred to by this `DocumentReference`.
@@ -1096,6 +1270,7 @@ declare namespace FirebaseFirestore {
      * @param moreFieldsOrPrecondition An alternating list of field paths and
      * values to update, optionally followed by a `Precondition` to enforce on
      * this update.
+     * @throws Error If the provided input is not valid Firestore data.
      * @return A Promise resolved with the write time of this update.
      */
     update(
@@ -1687,10 +1862,11 @@ declare namespace FirebaseFirestore {
      * it a document ID automatically.
      *
      * @param data An Object containing the data for the new document.
+     * @throws Error If the provided input is not a valid Firestore document.
      * @return A Promise resolved with a `DocumentReference` pointing to the
      * newly created document after it has been written to the backend.
      */
-    add(data: T): Promise<DocumentReference<T>>;
+    add(data: WithFieldValue<T>): Promise<DocumentReference<T>>;
 
     /**
      * Returns true if this `CollectionReference` is equal to the provided one.